+++ src/components/selection-create/SelectionCreateSpotItem.tsx
@@ -1,7 +1,6 @@
-import useModalExtraData from "@/hooks/useModalExtraData";
-import { IModalCreateSelectionSpotExtraData, ISelectionSpot } from "@/models/selection.model";
-import { useModalStore } from "@/stores/modalStore";
-import { useSelectionCreateStore, useSelectionSpotCreateStore } from "@/stores/selectionCreateStore";
+import useOpenSelectionSpotEditModal from "@/hooks/useOpenSelectionSpotEditModal";
+import { ISelectionSpot } from "@/models/selection.model";
+import { useSelectionCreateStore } from "@/stores/selectionCreateStore";
 import Image from "next/image";
 import { useStore } from "zustand";
 
@@ -17,27 +16,8 @@ const SelectionCreateSpotItem : React.FC<ISelectionCreateSpotItemProps> = ({
   index, 
   spotCategories
 }) => {
+  const openSpotEditModal = useOpenSelectionSpotEditModal();
   const { deleteSpot } = useStore(useSelectionCreateStore);
-  const { 
-    openModal,
-  } = useStore(useModalStore);
-  const { setExtraData } = useModalExtraData<IModalCreateSelectionSpotExtraData>();
-
-  const {
-    setPlaceName,
-    setAddress,
-    setCurrentCoordinate,
-    setTitle,
-    setDescription,
-    setCategory,
-    setSelectedLocation,
-    setSpotDescription,
-    setSpotPhoto,
-    setSpotPhoto1,
-    setSpotPhoto2,
-    setSpotPhoto3,
-    addHashtag,
-  } = useStore(useSelectionSpotCreateStore);
 
   const handleDeleteSpotClick = (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
@@ -46,54 +26,12 @@ const SelectionCreateSpotItem : React.FC<ISelectionCreateSpotItemProps> = ({
 
   const handleEditSpotClick = (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
-    setExtraData({
-      spotCategories: spotCategories,
-      index,
-    });
-    setPlaceName(spot.title);
-    setAddress(spot.formattedAddress);
-    setCurrentCoordinate({
-      lat: spot.latitude,
-      lng: spot.longitude,
-    });
-    setTitle(spot.title);
-    setDescription(spot.description);
-    setCategory({
-      id: spot.category,
-      name: spotCategories.find(
-        (category) => category.id === spot.category
-      )?.name || "",
-    });
-    setSelectedLocation({
-      key: 'User\'s current location',
-      placeId: spot.placeId,
-      address: spot.formattedAddress,
-      location: {
-        lat: spot.latitude,
-        lng: spot.longitude,
-      },
-    });
-    setSpotDescription(spot.description);
-
-    if (spot.photos.length > 0) {
-      setSpotPhoto(spot.photos[0]);
-    }
-    if (spot.photos.length > 1) {
-      setSpotPhoto1(spot.photos[1]);
-    }
-    if (spot.photos.length > 2) {
-      setSpotPhoto2(spot.photos[2]);
-    }
-    if (spot.photos.length > 3) {
-      setSpotPhoto3(spot.photos[3]);
-    }
-    
-    spot.hashtags.forEach((hashtag) => {
-      addHashtag(hashtag as string);
-    });
-    
-    openModal('GoogleMapsAddSelectionSpot');
-  }
+    openSpotEditModal(
+      spotCategories, 
+      spot, 
+      index
+    );
+  };
 
   return (
     <div key={index} className="flex items-center justify-between w-full">
